requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
openai==1.54.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.2
gunicorn==21.2.0
httpx==0.27.2